fix(Security): fix circular reference
JwtAuthenticationFilter → UserDetailsService(UserService)
           ↓
UserService → AuthenticationManager
           ↓
SecurityConfig → JwtAuthenticationFilter

Seperate AuthService so UserService wont refer AuthenticationManager
